requirements.txt
Flask==2.3.3
Flask-CORS==4.0.0
biopython==1.81
numpy==1.24.3
scikit-learn==1.3.0
werkzeug==2.3.7
python-dateutil==2.8.2